requirements.txt
spotipy
Flask
python-dotenv
